components folder rename to pages and About page added
