+++ frontend/src/pages/AboutPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const AboutPage = () => {
+  return (
+    <div>
+      About Page
+    </div>
+  )
+}
+
+export default AboutPage
